Removes remaining CASinoCore references
For some reason, the RSpec :around block that the controller specs were
using to mount the dummy application's routes was causing specs to pass
without actually running the spec. This is most likely an RSpec bug,
but I didn't have time to dig too deep into that.

The problem is that those passing tests were hiding several more
CASinoCore references.

* Works around the RSpec issue by using the more formally
supported `use_route: :casino` parameter that is passed in the spec.

* Removes all remaining CASinoCore references that were obfuscated by
  improperly passing specs

* Fixes a few specs that were passing the HTTP parameters as a nested
  hash instead of a single-level one that were also obfuscated by
  improperly passing specs

* Removes double mounting of CASino in dummy application